requirements.txt
agate==1.11.0
agate-dbf==0.2.3
agate-excel==0.4.1
agate-sql==0.7.2
appdirs==1.4.4
asgiref==3.8.1
Babel==2.15.0
certifi==2024.2.2
chardet==4.0.0
charset-normalizer==3.3.2
csvkit==2.0.0
dbfread==2.0.7
Django==5.0.3
django-admin-honeypot==1.1.0
django-admin-honeypot-advanced==1.0.1
django-admin-thumbnails==0.2.5
django-ipware==7.0.1
django-paypal==2.1
django-session-timeout==0.1.0
et-xmlfile==1.1.0
fs==2.4.16
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
isodate==0.6.1
leather==0.4.0
olefile==0.47
openpyxl==3.1.4
packaging==24.0
parsedatetime==2.6
pillow==10.2.0
pipdeptree==2.23.0
psycopg2-binary==2.9.9
python-decouple==3.8
python-ipware==3.0.0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2024.1
razorpay==1.4.2
requests==2.31.0
simplejson==3.19.2
six==1.15.0
SQLAlchemy==2.0.31
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
xlrd==2.0.1
